+++ persistent/Database/Persist/Class/PersistField.hs
@@ -1,4 +1,5 @@
 {-# LANGUAGE CPP #-}
+{-# LANGUAGE GADTs #-}
 {-# LANGUAGE ExistentialQuantification #-}
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE PatternGuards, DataKinds, TypeOperators, UndecidableInstances #-}
@@ -450,10 +451,15 @@ getPersistMap (PersistByteString bs)
 getPersistMap PersistNull = Right []
 getPersistMap x = Left $ fromPersistValueError "[(Text, PersistValue)]" "map, string, bytestring or null" x
 
-data SomePersistField = forall a. PersistField a => SomePersistField a
+data SomePersistField where
+    SomePersistField
+        :: (PersistField a)
+        => a
+        -> SomePersistField
+
 instance PersistField SomePersistField where
     toPersistValue (SomePersistField a) = toPersistValue a
-    fromPersistValue x = fmap SomePersistField (fromPersistValue x :: Either Text Text)
+    fromPersistValue x = error "oh no" $ fmap SomePersistField (fromPersistValue x :: Either Text Text)
 
 instance PersistField Checkmark where
     toPersistValue Active   = PersistBool True